+++ src/components/o-progress-bar/js/log-level.js
